Started to segment off paleogeography subsystem
