+++ cli/macrostrat/cli/subsystems/paleogeography/corelle/fixtures/02-carto-slim-rotated.sql
