+++ cli/macrostrat/cli/subsystems/paleogeography/corelle/sql/carto-plate-index-cache.sql
